rest-service/requirements.txt
#
# This file is autogenerated by pip-compile with python 3.10
# To update, run:
#
#    pip-compile --output-file=requirements.txt requirements.in setup.py
#
aiohttp==3.8.1
    # via cloudify-common
aiosignal==1.2.0
    # via aiohttp
alembic==1.8.1
    # via flask-migrate
aniso8601==9.0.1
    # via flask-restful
async-timeout==4.0.2
    # via aiohttp
attrs==22.1.0
    # via
    #   aiohttp
    #   jsonschema
babel==2.10.3
    # via flask-babelex
blinker==1.5
    # via
    #   flask-mail
    #   flask-principal
bottle==0.12.23
    # via cloudify-common
cachetools==3.1.1
    # via cloudify-rest-service (setup.py)
certifi==2022.6.15
    # via requests
cffi==1.14.6
    # via
    #   cloudify-rest-service (setup.py)
    #   cryptography
charset-normalizer==2.1.1
    # via
    #   aiohttp
    #   requests
click==8.1.3
    # via flask
cloudify-common @ https://github.com/cloudify-cosmo/cloudify-common/archive/flask-restful-1.zip
    # via
    #   -r requirements.in
    #   cloudify-rest-service (setup.py)
cryptography==38.0.1
    # via cloudify-rest-service (setup.py)
distro==1.7.0
    # via
    #   cloudify-common
    #   cloudify-rest-service (setup.py)
dnspython==2.2.1
    # via email-validator
email-validator==1.2.1
    # via cloudify-rest-service (setup.py)
fasteners==0.17.3
    # via cloudify-common
flask==2.2.2
    # via
    #   cloudify-rest-service (setup.py)
    #   flask-babelex
    #   flask-login
    #   flask-mail
    #   flask-migrate
    #   flask-principal
    #   flask-restful
    #   flask-security
    #   flask-sqlalchemy
    #   flask-wtf
flask-babelex==0.9.4
    # via flask-security
flask-login==0.6.2
    # via flask-security
flask-mail==0.9.1
    # via flask-security
flask-migrate==3.1.0
    # via cloudify-rest-service (setup.py)
flask-principal==0.4.0
    # via flask-security
flask-restful==0.3.9
    # via cloudify-rest-service (setup.py)
flask-security==3.0.0
    # via cloudify-rest-service (setup.py)
flask-sqlalchemy==2.5.1
    # via
    #   cloudify-rest-service (setup.py)
    #   flask-migrate
flask-wtf==1.0.1
    # via flask-security
frozenlist==1.3.1
    # via
    #   aiohttp
    #   aiosignal
greenlet==1.1.3
    # via sqlalchemy
idna==3.3
    # via
    #   email-validator
    #   requests
    #   yarl
itsdangerous==2.1.2
    # via
    #   flask
    #   flask-security
    #   flask-wtf
jinja2==3.1.2
    # via
    #   cloudify-common
    #   flask
    #   flask-babelex
jsonschema==3.2.0
    # via cloudify-rest-service (setup.py)
mako==1.2.2
    # via alembic
markupsafe==2.1.1
    # via
    #   jinja2
    #   mako
    #   werkzeug
    #   wtforms
multidict==6.0.2
    # via
    #   aiohttp
    #   yarl
networkx==2.8.8
    # via cloudify-common
packaging==21.3
    # via cloudify-rest-service (setup.py)
passlib==1.7.4
    # via flask-security
pika==1.3.0
    # via
    #   cloudify-common
    #   cloudify-rest-service (setup.py)
proxy-tools==0.1.0
    # via cloudify-common
psutil==5.9.2
    # via cloudify-rest-service (setup.py)
psycopg2==2.9.3
    # via cloudify-rest-service (setup.py)
pycparser==2.21
    # via cffi
pydantic==1.10.2
    # via cloudify-rest-service (setup.py)
pyparsing==3.0.9
    # via packaging
pyrsistent==0.18.1
    # via jsonschema
python-dateutil==2.8.2
    # via cloudify-rest-service (setup.py)
pytz==2022.2.1
    # via
    #   babel
    #   cloudify-common
    #   cloudify-rest-service (setup.py)
    #   flask-restful
pyyaml==6.0
    # via
    #   cloudify-common
    #   cloudify-rest-service (setup.py)
requests==2.28.1
    # via
    #   cloudify-common
    #   cloudify-rest-service (setup.py)
    #   requests-toolbelt
requests-toolbelt==0.9.1
    # via cloudify-common
retrying==1.3.3
    # via cloudify-rest-service (setup.py)
six==1.16.0
    # via
    #   flask-restful
    #   jsonschema
    #   python-dateutil
    #   retrying
speaklater==1.3
    # via flask-babelex
sqlalchemy==1.4.41
    # via
    #   alembic
    #   cloudify-rest-service (setup.py)
    #   flask-sqlalchemy
typing-extensions==4.3.0
    # via pydantic
urllib3==1.26.12
    # via requests
wagon==0.12.0
    # via
    #   cloudify-common
    #   cloudify-rest-service (setup.py)
werkzeug==2.2.2
    # via
    #   flask
    #   flask-login
wheel==0.38.1
    # via wagon
wtforms==3.0.1
    # via flask-wtf
yarl==1.8.1
    # via aiohttp

# The following packages are considered to be unsafe in a requirements file:
# setuptools
